# app/adapters/inbound/web/routes.py
"""FastAPI routes for the web UI and API."""

import asyncio
import contextlib
import json
import logging
import os
from uuid import UUID

from fastapi import APIRouter, HTTPException, Request, UploadFile
from fastapi.responses import FileResponse, PlainTextResponse, Response, StreamingResponse
from fastapi.templating import Jinja2Templates

from app.adapters.inbound.web.schemas import (
    AudioFileSchema,
    HealthResponse,
    JobStatusResponse,
    TranscriptionResultResponse,
    UploadResponse,
    UrlUploadRequest,
)
from app.application.dto import SubmitTranscriptionRequest, SubmitUrlTranscriptionRequest
from app.bootstrap import get_container
from app.domain.exceptions import (
    DomainError,
    FileTooLargeError,
    InvalidAudioFormatError,
    InvalidStateTransitionError,
    InvalidUrlError,
    MaxRetriesExceededError,
)

logger = logging.getLogger(__name__)

router = APIRouter()

_templates_dir = os.path.join(os.path.dirname(__file__), "templates")
templates = Jinja2Templates(directory=_templates_dir)


# ── HTML Pages ──────────────────────────────────────────────


@router.get("/")
async def upload_page(request: Request):
    return templates.TemplateResponse("upload.html", {"request": request})


@router.get("/jobs/{job_id}")
async def job_page(request: Request, job_id: UUID):
    container = get_container()
    job = container.get_job_status.execute(job_id)
    if job is None:
        raise HTTPException(status_code=404, detail="Job not found")

    result = container.get_job_status.get_result(job_id)

    return templates.TemplateResponse(
        "job.html",
        {"request": request, "job": job, "result": result},
    )


# ── API Endpoints ───────────────────────────────────────────


@router.post("/api/upload", status_code=201, response_model=UploadResponse)
async def upload_file(file: UploadFile, language: str = "pt-BR"):
    container = get_container()

    file_data = await file.read()

    try:
        request = SubmitTranscriptionRequest(
            filename=file.filename or "unknown",
            file_data=file_data,
            language=language,
        )
        response = container.submit_transcription.execute(request)
    except InvalidAudioFormatError as e:
        raise HTTPException(status_code=400, detail=str(e)) from e
    except FileTooLargeError as e:
        raise HTTPException(status_code=413, detail=str(e)) from e
    except Exception as e:
        logger.error(f"Upload failed: {e}")
        raise HTTPException(status_code=503, detail="Service unavailable") from e

    return UploadResponse(
        job_id=response.job_id,
        status=response.status,
        redirect_url=response.redirect_url,
    )


@router.post("/api/upload-url", status_code=201, response_model=UploadResponse)
async def upload_from_url(body: UrlUploadRequest):
    container = get_container()

    try:
        request = SubmitUrlTranscriptionRequest(
            url=body.url,
            language=body.language,
        )
        response = container.submit_transcription.execute_from_url(request)
    except InvalidUrlError as e:
        raise HTTPException(status_code=422, detail=str(e)) from e
    except DomainError as e:
        raise HTTPException(status_code=400, detail=str(e)) from e
    except Exception as e:
        logger.error(f"URL upload failed: {e}")
        raise HTTPException(status_code=503, detail="Service unavailable") from e

    return UploadResponse(
        job_id=response.job_id,
        status=response.status,
        redirect_url=response.redirect_url,
    )


@router.get("/api/jobs")
async def list_jobs(limit: int = 50, offset: int = 0):
    container = get_container()
    jobs = container.repository.get_all_jobs(limit=limit, offset=offset)
    result = []
    for job in jobs:
        audio = container.repository.get_audio_file(job.audio_file_id)
        result.append(
            {
                "job_id": str(job.id),
                "status": job.status.value,
                "progress_percent": job.progress_percent,
                "language": job.language,
                "engine_name": job.engine_name,
                "original_filename": audio.original_filename if audio else "unknown",
                "created_at": job.created_at.strftime("%Y-%m-%dT%H:%M:%SZ"),
                "error_message": job.error_message,
                "size_bytes": audio.size_bytes if audio else 0,
                "duration_seconds": audio.duration_seconds if audio else None,
                "format": audio.format.value if audio else None,
                "source_url": audio.source_url if audio else None,
            }
        )
    return result


@router.delete("/api/jobs")
async def delete_all_jobs():
    """Delete all transcription jobs, results, and associated audio files."""
    container = get_container()

    jobs = container.repository.get_all_jobs(limit=1000)
    for job in jobs:
        audio = container.repository.get_audio_file(job.audio_file_id)
        if audio:
            with contextlib.suppress(Exception):
                container.storage.delete(audio.storage_path)
            if audio.converted_path:
                with contextlib.suppress(Exception):
                    container.storage.delete(audio.converted_path)

    count = container.repository.delete_all_jobs()
    logger.info(f"Deleted {count} transcription jobs and associated files")
    return {"deleted": count}


@router.post("/api/jobs/{job_id}/retry")
async def retry_job(job_id: UUID):
    """Retry a failed transcription job."""
    container = get_container()

    job = container.repository.get_job(job_id)
    if job is None:
        raise HTTPException(status_code=404, detail="Job not found")

    try:
        job.retry()
        container.repository.save_job(job)
        container.queue.enqueue(job.id)
    except MaxRetriesExceededError:
        raise HTTPException(status_code=409, detail="Maximum retries exceeded") from None
    except InvalidStateTransitionError as e:
        raise HTTPException(status_code=409, detail=str(e)) from e

    logger.info(f"Retrying job {job_id} (attempt {job.retry_count}/3)")
    return {
        "job_id": str(job.id),
        "status": job.status.value,
        "retry_count": job.retry_count,
    }


@router.get("/api/jobs/{job_id}", response_model=JobStatusResponse)
async def get_job_status(job_id: UUID):
    container = get_container()
    job = container.get_job_status.execute(job_id)
    if job is None:
        raise HTTPException(status_code=404, detail="Job not found")

    audio_file_schema = None
    if job.audio_file:
        audio_file_schema = AudioFileSchema(
            original_filename=job.audio_file.original_filename,
            format=job.audio_file.format,
            size_bytes=job.audio_file.size_bytes,
            duration_seconds=job.audio_file.duration_seconds,
            source_url=job.audio_file.source_url,
        )

    return JobStatusResponse(
        job_id=job.job_id,
        status=job.status,
        progress_percent=job.progress_percent,
        language=job.language,
        engine_name=job.engine_name,
        audio_file=audio_file_schema,
        created_at=job.created_at,
        updated_at=job.updated_at,
        error_message=job.error_message,
    )


@router.get("/api/jobs/{job_id}/result", response_model=TranscriptionResultResponse)
async def get_result(job_id: UUID):
    container = get_container()
    result = container.get_job_status.get_result(job_id)
    if result is None:
        raise HTTPException(status_code=404, detail="Job not found or not yet completed")

    return TranscriptionResultResponse(
        job_id=result.job_id,
        full_text=result.full_text,
        language=result.language,
        engine_name=result.engine_name,
        processing_duration_seconds=result.processing_duration_seconds,
    )


@router.get("/api/jobs/{job_id}/result/download")
async def download_result(job_id: UUID):
    container = get_container()
    result = container.get_job_status.get_result(job_id)
    if result is None:
        raise HTTPException(status_code=404, detail="Job not found or not yet completed")

    job = container.get_job_status.execute(job_id)
    filename = "transcription.txt"
    if job and job.audio_file:
        base = os.path.splitext(job.audio_file.original_filename)[0]
        filename = f"{base}.txt"

    return PlainTextResponse(
        content=result.full_text,
        headers={"Content-Disposition": f'attachment; filename="{filename}"'},
    )


@router.get("/api/jobs/{job_id}/result/download/{fmt}")
async def download_result_format(job_id: UUID, fmt: str):
    """Download transcription in TXT, SRT, or VTT format."""
    if fmt not in ("txt", "srt", "vtt"):
        raise HTTPException(status_code=400, detail=f"Unsupported format: {fmt}. Use txt, srt, or vtt.")

    container = get_container()
    result = container.get_job_status.get_result(job_id)
    if result is None:
        raise HTTPException(status_code=404, detail="Job not found or not yet completed")

    job = container.get_job_status.execute(job_id)
    base = "transcription"
    duration = 0.0
    if job and job.audio_file:
        base = os.path.splitext(job.audio_file.original_filename)[0]
        duration = job.audio_file.duration_seconds or 0.0

    if fmt == "txt":
        return PlainTextResponse(
            content=result.full_text,
            headers={"Content-Disposition": f'attachment; filename="{base}.txt"'},
        )

    from app.domain.services.subtitle_generator import generate_srt, generate_vtt

    if fmt == "srt":
        content = generate_srt(result.full_text, duration)
        return Response(
            content=content,
            media_type="application/x-subrip",
            headers={"Content-Disposition": f'attachment; filename="{base}.srt"'},
        )

    content = generate_vtt(result.full_text, duration)
    return Response(
        content=content,
        media_type="text/vtt",
        headers={"Content-Disposition": f'attachment; filename="{base}.vtt"'},
    )


@router.get("/api/jobs/{job_id}/audio")
async def stream_audio(job_id: UUID):
    """Stream audio file for in-browser playback with Range request support."""
    container = get_container()

    job = container.repository.get_job(job_id)
    if job is None:
        raise HTTPException(status_code=404, detail="Job not found")

    audio = container.repository.get_audio_file(job.audio_file_id)
    if audio is None:
        raise HTTPException(status_code=404, detail="Audio file not found")

    # Prefer converted WAV (better for playback), fall back to original
    path = audio.converted_path or audio.storage_path
    if not path:
        raise HTTPException(status_code=404, detail="Audio file not available")

    absolute_path = container.storage.get_absolute_path(path)
    if not os.path.exists(absolute_path):
        raise HTTPException(status_code=404, detail="Audio file not found on disk")

    # Determine content type
    ext = os.path.splitext(absolute_path)[1].lower()
    content_types = {
        ".wav": "audio/wav",
        ".mp3": "audio/mpeg",
        ".ogg": "audio/ogg",
        ".flac": "audio/flac",
        ".m4a": "audio/mp4",
    }
    media_type = content_types.get(ext, "application/octet-stream")

    return FileResponse(absolute_path, media_type=media_type, filename=os.path.basename(absolute_path))


@router.post("/api/sample", status_code=201, response_model=UploadResponse)
async def submit_sample(language: str = "pt-BR"):
    """Submit the bundled sample audio for demo transcription."""
    base = os.path.dirname(__file__)
    sample_path = os.path.normpath(os.path.join(base, "..", "..", "..", "..", "static", "sample", "sample-pt-br.wav"))

    if not os.path.exists(sample_path):
        raise HTTPException(status_code=503, detail="Sample audio not available")

    with open(sample_path, "rb") as f:
        file_data = f.read()

    container = get_container()

    try:
        request = SubmitTranscriptionRequest(
            filename="sample-pt-br.wav",
            file_data=file_data,
            language=language,
        )
        response = container.submit_transcription.execute(request)
    except Exception as e:
        logger.error(f"Sample submission failed: {e}")
        raise HTTPException(status_code=503, detail="Service unavailable") from e

    return UploadResponse(
        job_id=response.job_id,
        status=response.status,
        redirect_url=response.redirect_url,
    )


@router.get("/api/jobs/{job_id}/progress")
async def job_progress_sse(job_id: UUID):
    container = get_container()

    async def event_stream():
        last_status = None
        last_progress = -1

        while True:
            job = container.get_job_status.execute(job_id)
            if job is None:
                yield 'event: error\ndata: {"error": "Job not found"}\n\n'
                return

            if job.status != last_status or job.progress_percent != last_progress:
                last_status = job.status
                last_progress = job.progress_percent
                data = json.dumps({"status": job.status, "progress_percent": job.progress_percent})
                yield f"event: status\ndata: {data}\n\n"

            if job.status in ("COMPLETED", "FAILED"):
                done_data = json.dumps({"redirect_url": f"/jobs/{job_id}"})
                yield f"event: done\ndata: {done_data}\n\n"
                return

            await asyncio.sleep(1)

    return StreamingResponse(
        event_stream(),
        media_type="text/event-stream",
        headers={
            "Cache-Control": "no-cache",
            "Connection": "keep-alive",
            "X-Accel-Buffering": "no",
        },
    )


@router.get("/health")
async def health():
    """Simple health check for Dokku zero-downtime deploys."""
    return {"status": "ok", "service": "Voxscribe"}


@router.get("/api/health", response_model=HealthResponse)
async def health_check():
    container = get_container()

    redis_status = "disconnected"
    try:
        from redis import Redis

        r = Redis.from_url(container.settings.redis_url)
        r.ping()
        redis_status = "connected"
    except Exception:
        pass

    return HealthResponse(
        status="healthy",
        engine=container.settings.transcription_engine,
        redis=redis_status,
    )